tool/requirements.txt
dash==2.6.2
jinja2==3.1.3
pandas==1.5.3
Pillow==10.4.0
plotly==5.20.0
scipy==1.12.0
setuptools==69.2.0
kaleido==0.2.1
boto3==1.35.49
numpy==1.26.4